+++ src/main/java/my_mall/entity/vo/OrderDetailVO.java
@@ -15,9 +15,9 @@ public class OrderDetailVO {
     private  String orderNo;
     private  Long userId;
     private  Integer totalPrice;
-    private  Byte payStatus;
-    private  Byte payType;
-    private  Byte orderStatus;
+    private  Integer payStatus;
+    private  Integer payType;
+    private  Integer orderStatus;
     private LocalDateTime payTime;
     private  LocalDateTime createTime;
     private List<OrderCartDTO> orderCartDTO;
